refactor: modularized patients table
