+++ src/components/pages/ListPatients.js
@@ -3,8 +3,7 @@ import { useLocation } from 'react-router-dom'
 import { useState, useEffect } from 'react'
 import Message from '../layout/Message'
 import Container from '../layout/Container'
-import { DataGrid } from '@material-ui/data-grid'
-
+import TablePatients from '../patients/TablePatients';
 
 import styles from './ListPatients.module.css'
 
@@ -38,42 +37,13 @@ export default function ListPatients() {
 
     }, [])
 
-    const columns = [
-        { field: 'id', headerName: 'ID', width: 100 },
-        { field: 'name', headerName: 'Nome', width: 180 },
-        { field: 'cpf', headerName: 'CPF', width: 130 },
-        { field: 'gender', headerName: 'Sexo', width: 130 },
-        {
-          field: 'date',
-          headerName: 'Data de nascimento',
-          type: 'date',
-          width: 200,
-        },
-
-        { field: 'status', headerName: 'Status', width: 120 },
-
-      ];
+    
 
     return (
         <div className={ styles.ListPatientContainer }>
             <h1>Pacientes</h1>
 
-          
-
-
-
-
-            <div style={{ height: 400, width: '900px' }}>
-            <DataGrid
-                rows={rows}
-                columns={columns}
-                pageSize={5}
-                rowsPerPageOptions={[5]}
-            />
-            </div>
-
-            
-            
+            <TablePatients rows={ rows }/>
 
             { message && (
                 <Message message={ message } type="success" />
